Clean up MardownRender to have both body and comment
