+++ app/components/Issues/Add/Verify/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import T from 'prop-types';
 
 import { LanguageWrapper } from 'components/base_ui';
-import { CommentCard } from 'components/MarkdownRender';
+import { BodyCard } from 'components/MarkdownRender';
 import { issueDataDictionary } from 'containers/Issues/constants';
 
 import {
@@ -19,17 +19,8 @@ export class VerifyForm extends React.PureComponent {
   render() {
     const {
       issueData: { issueBody, issueLanguages, issueName, issueUrl },
-      activeUser: { id, profilePic, username },
-      handleNav,
     } = this.props;
 
-    const primaryUser = {
-      small: true,
-      detailRoute: `/users/detail/${id}`,
-      alt: username,
-      username,
-      profilePic,
-    };
     const mapLanguages = array => {
       if (array.value.length > 0) {
         return array.value.map(el => (
@@ -48,12 +39,7 @@ export class VerifyForm extends React.PureComponent {
         </ValueWrapper>
         <StyledLink>{issueUrl.value}</StyledLink>
         <ValueWrapper>
-          <CommentCard
-            body={issueBody.value}
-            date={Date.now()}
-            handleNav={handleNav}
-            userProfile={primaryUser}
-          />
+          <BodyCard body={issueBody.value} />
         </ValueWrapper>
         <ValueWrapper>
           <Divider />
@@ -67,10 +53,6 @@ export class VerifyForm extends React.PureComponent {
   }
 }
 
-VerifyForm.propTypes = {
-  activeUser: T.object,
-  issueData: T.object,
-  handleNav: T.func,
-};
+VerifyForm.propTypes = { issueData: T.object };
 
 export default VerifyForm;
